Add files via upload
In this game typically a user and the computer, compete against each other in a series of rounds. Each player chooses one of three options: Snake, Water, or Gun. The outcome of each round is determined by comparing the choices made by the players:

1.If both players choose the same option, it's a tie.
2.Snake beats Water (because it can drink the water).
3.Water beats Gun (because it can rust the gun).
4.Gun beats Snake (because it can shoot the snake).

The game continues for a predefined number of rounds, and the player with the highest score at the end wins the game.🏆🏆